Disable mouse input when we dont have map available
