+++ src/components/Map/data.ts
@@ -1,0 +1,25 @@
+type Source = string;
+export const MapData: Record<string, Source> = {
+  callisto: 'callisto.jpg',
+  dione: 'dione.jpg',
+  earth: 'earth_bluemarble.jpg',
+  enceladus: 'enceladus.jpg',
+  europa: 'europa.jpg',
+  ganymede: 'ganymede.jpg',
+  iapetus: 'iapetus.jpg',
+  io: 'io.jpg',
+  jupiter: 'jupiter.jpg',
+  mars: 'mars.jpg',
+  mercury: 'mercury.jpg',
+  mimas: 'mimas.jpg',
+  moon: 'moon.jpg',
+  neptune: 'neptune.jpg',
+  phobos: 'phobos.jpg',
+  rhea: 'rhea.jpg',
+  saturn: 'saturn.jpg',
+  tethys: 'tethys.jpg',
+  titan: 'titan.jpg',
+  triton: 'triton.jpg',
+  uranus: 'uranus.jpg',
+  venus: 'venus.jpg'
+};
